+++ WpfApplication1/Models/Tara.cs
@@ -12,8 +12,8 @@ namespace WpfApplication1
         public Guid Id { get; set; }
         public string RFID_Metka { get; set; }
         public string Name { get; set; }
-        public int WeightNetto { get; set; }
-        public int Volume { get; set; }
+        public ushort WeightNetto { get; set; }
+        public ushort Volume { get; set; }
         public Guid CoctailId { get; set; }
 
         public Tara()
